# src/routes/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Darve Server API",
    "description": "API documentation for the Darve Server including users, posts, notifications, auth, discussions, tasks, wallet, and follows",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "",
      "description": "Development server"
    }
  ],
  "security": [],
  "paths": {
...
    "/api/users/{user_id}/followers/count": {
      "get": {
        "tags": ["Follows"],
        "summary": "Get user's followers count",
        "description": "Get the number of followers for a specific user",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "description": "The ID of the user",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Followers count",
            "content": {
              "application/json": {
                "schema": {
                  "type": "integer",
                  "format": "int32",
                  "description": "Number of followers"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/api/users/{user_id}/following/count": {
      "get": {
        "tags": ["Follows"],
        "summary": "Get user's following count",
        "description": "Get the number of users a specific user is following",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "description": "The ID of the user",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Following count",
            "content": {
              "application/json": {
                "schema": {
                  "type": "integer",
                  "format": "int32",
                  "description": "Number of users being followed"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
...
    "/api/users/{user_id}/followers": {
      "get": {
        "tags": ["Follows"],
        "summary": "Get user followers",
        "description": "Get followers of a specific user",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "description": "ID of the user",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of followers",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/UserItemView"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/api/users/{user_id}/following": {
      "get": {
        "tags": ["Follows"],
        "summary": "Get users followed by user",
        "description": "Get users that a specific user is following",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "description": "ID of the user",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of followed users",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/UserItemView"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/api/followers/{follow_user_id}": {
      "post": {
        "tags": ["Follows"],
        "summary": "Follow user",
        "description": "Follow a specific user",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "parameters": [
          {
            "name": "follow_user_id",
            "in": "path",
            "required": true,
            "description": "ID of the user to follow",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User followed successfully"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "User not found"
          }
        }
      },
      "delete": {
        "tags": ["Follows"],
        "summary": "Unfollow user",
        "description": "Unfollow a specific user",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "parameters": [
          {
            "name": "follow_user_id",
            "in": "path",
            "required": true,
            "description": "ID of the user to unfollow",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User unfollowed successfully"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "User not found"
          }
        }
      },
      "get": {
        "tags": ["Follows"],
        "summary": "Check if following user",
        "description": "Check if current user is following a specific user",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "parameters": [
          {
            "name": "follow_user_id",
            "in": "path",
            "required": true,
            "description": "ID of the user to check",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Following status checked successfully"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "User not found"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "UserItemView": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "User ID"
          },
          "username": {
            "type": "string",
            "description": "Username"
          },
          "name": {
            "type": "string",
            "description": "Full name"
          },
          "image_url": {
            "type": "string",
            "description": "URL of the user's profile image"
          }
        }
      },
...
    },
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "jwt"
      }
    }
  },
...
}